MAIS/requirements.txt
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.2.0
PyQt6==6.6.1
PyQt6_sip==13.6.0
rawpy==0.19.0
